change build structure in makefile
The makefile used to prepend a _ to the .o/.s file generated from a .c
file. This was so that if the user had a foo.c, foo.s, and foo.h, the
.s and .o file generated from compiling and assembling the .c file
wouldn't conflict with foo.s and foo.o generated by assembling foo.s.

Now instead of doing that, build product of .c and .s are kept separate,
so there's no danger of naming collisions.

Additionally, the makefile searches for .c and .s files in the
appropriate directory instead of having to list them manually.

Finally, dependencies for .c and .s files are generated by the compiler,
rather than being forgotten about.